add test for toggle button
