+++ ui/package.json
@@ -40,6 +40,7 @@
     "autoprefixer": "^10.4.7",
     "babel-eslint": "^10.1.0",
     "postcss": "^8.4.14",
+    "sinon": "^14.0.0",
     "tailwindcss": "^3.0.24"
   }
 }
